Fix `about.hpp` typo and refactor
